requirements.txt
langchain==0.2.3
langchain_community==0.2.4
langchain_core==0.2.5
langchain_google_community==1.0.5
langchain_openai==0.1.8
Pillow==10.3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.35.0
transformers==4.41.2
